Break apart roadway change tests; add asserts to managed lane tests
- To better manage and run them, tests for different apply methods are separated in different files
- managed lane tests now have explicit assert statements testing actual truthiness
Also:
- convenience method added to RoadwayNetwork to count number of managed lane links
