--- tests/test_roadway/test_changes/test_changes.py
@@ -1,826 +1,0 @@
-import copy
-import os
-
-import pytest
-
-import pandas as pd
-
-from network_wrangler import ProjectCard
-from network_wrangler import RoadwayNetwork
-from network_wrangler import WranglerLogger
-
-from network_wrangler.roadway.model_roadway import (
-    MANAGED_LANES_LINK_ID_SCALAR,
-)
-
-pd.set_option("display.max_rows", 500)
-pd.set_option("display.max_columns", 500)
-pd.set_option("display.width", 50000)
-
-"""
-Run just the tests labeled basic using `pytest tests/test_roadway/test_changes/test_changes.py`
-To run with print statments, use `pytest -s tests/test_roadway/test_changes/test_changes.py`
-"""
-
-def test_change_roadway_existing_and_change_single_link(request,stpaul_net):
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    
-    #Set facility selection
-    _facility = {
-        "links":[{"osm_link_id": ['223371529']}],
-        "A": {"osm_node_id": '187854529'}, # Jackson St
-        "B": {"osm_node_id": '187899923'},  # Robert St N    
-    }
-    _properties = [{
-        'property':'lanes',  #changes number of lanes 3 to 2 (reduction of 1)
-        'existing': 2,
-        'change': -1
-    }]
-
-    _selected_link_idx = net.select_roadway_features(_facility)
-    _p_to_track = ["name"]+ [p["property"] for p in _properties ]
-
-    WranglerLogger.debug(f"_p_to_track: {_p_to_track}")
-
-    _orig_links = pd.DataFrame(copy.deepcopy(net.links_df))
-    _orig_links = _orig_links.loc[_selected_link_idx,_p_to_track]
-    WranglerLogger.debug(f"_orig_links:\n{_orig_links}")
-
-    # apply change
-    net = net.apply_roadway_feature_change(_selected_link_idx, _properties)
-
-    _rev_links = pd.DataFrame(net.links_df)
-    _rev_links = _rev_links.loc[_selected_link_idx,_p_to_track]
-    WranglerLogger.debug(f"_rev_links:\n{_rev_links}")
-    
-    WranglerLogger.debug(f"ORIGINAL to REVISED Comparison\n {_orig_links.compare(_rev_links)}")
-
-    for p in _properties:
-        _expected_value = p['existing']+p['change']
-        WranglerLogger.debug(f"Expected_value of {p['property']}:{_expected_value}")
-        assert _rev_links[p['property']].eq(_expected_value).all()
-
-def test_change_multiple_properties_multiple_links(request,stpaul_net):
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    #Set facility selection
-    _facility = {
-        "links":[{"name": ['6th','Sixth','sixth']}],
-        "A": {"osm_node_id": '187899923'}, # Jackson St
-        "B": {"osm_node_id": '187865924'},  # Robert St N    
-    }
-    _properties = [
-        {
-            'property':'lanes', 
-            'set': 2,
-        },
-        {
-            'property':'bus_only', 
-            'set': 1,
-        },
-        {
-            'property':'drive_access', 
-            'set': 0,
-        },
-    ]
-
-    _selected_link_idx = net.select_roadway_features(_facility)
-    _p_to_track = ["name"]+ [p["property"] for p in _properties ]
-
-    WranglerLogger.debug(f"_p_to_track: {_p_to_track}")
-
-    _orig_links = pd.DataFrame(copy.deepcopy(net.links_df))
-    _orig_links = _orig_links.loc[_selected_link_idx,_p_to_track]
-    WranglerLogger.debug(f"_orig_links:\n{_orig_links}")
-
-    # apply change
-    net = net.apply_roadway_feature_change(_selected_link_idx, _properties)
-
-    _rev_links = pd.DataFrame(net.links_df)
-    _rev_links = _rev_links.loc[_selected_link_idx,_p_to_track]
-    WranglerLogger.debug(f"_rev_links:\n{_rev_links}")
-    
-    WranglerLogger.debug(f"ORIGINAL to REVISED Comparison\n {_orig_links.compare(_rev_links)}")
-
-    for p in _properties:
-        _expected_value = p['set']
-        WranglerLogger.debug(f"Expected_value of {p['property']}:{_expected_value}")
-        assert _rev_links[p['property']].eq(_expected_value).all()
-
-
-def test_change_multiple_properties_multiple_links_existing_set(request,stpaul_net):
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    #Set facility selection
-    _facility = {
-        "links":[{"name": ['6th','Sixth','sixth']}],
-        "A": {"osm_node_id": '187899923'}, # Jackson St
-        "B": {"osm_node_id": '187865924'},  # Robert St N    
-    }
-    _properties = [
-        {
-            'property':'lanes', 
-            'existing':1,
-            'set': 2,
-        },
-        {
-            'property':'bus_only', 
-            'set': 1,
-        },
-        {
-            'property':'drive_access', 
-            'set': 0,
-        },
-    ]
-
-    _selected_link_idx = net.select_roadway_features(_facility)
-    _p_to_track = ["name"]+ [p["property"] for p in _properties ]
-
-    WranglerLogger.debug(f"_p_to_track: {_p_to_track}")
-
-    _orig_links = pd.DataFrame(copy.deepcopy(net.links_df))
-    _orig_links = _orig_links.loc[_selected_link_idx,_p_to_track]
-    WranglerLogger.debug(f"_orig_links:\n{_orig_links}")
-
-    # apply change
-    net = net.apply_roadway_feature_change(_selected_link_idx, _properties)
-
-    _rev_links = pd.DataFrame(net.links_df)
-    _rev_links = _rev_links.loc[_selected_link_idx,_p_to_track]
-    WranglerLogger.debug(f"_rev_links:\n{_rev_links}")
-    
-    WranglerLogger.debug(f"ORIGINAL to REVISED Comparison\n {_orig_links.compare(_rev_links)}")
-
-    for p in _properties:
-        _expected_value = p['set']
-        WranglerLogger.debug(f"Expected_value of {p['property']}:{_expected_value}")
-        assert _rev_links[p['property']].eq(_expected_value).all()
-
-@pytest.mark.menow
-def test_add_managed_lane(request, stpaul_net, stpaul_ex_dir, scratch_dir):
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    print("Reading project card ...")
-    project_card_name = "5_managed_lane.yml"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path)
-
-    print("Selecting roadway features ...")
-    selected_link_indices = net.select_roadway_features(project_card.facility)
-
-    attributes_to_update = [p["property"] for p in project_card.properties]
-    orig_links = net.links_df.loc[
-        selected_link_indices, net.links_df.columns.intersection(attributes_to_update)
-    ]
-    print("Original Links:\n", orig_links)
-
-    net = net.apply_managed_lane_feature_change(
-        net.select_roadway_features(project_card.facility), project_card.properties
-    )
-
-    rev_links = net.links_df.loc[selected_link_indices, attributes_to_update]
-    print("Revised Links:\n", rev_links)
-
-    net.write(filename="test_ml", model=True, path=scratch_dir)
-
-    print("--Finished:", request.node.name)
-
-def test_add_adhoc_field(request, stpaul_net):
-    """
-    Makes sure new fields can be added in the API and be saved and read in again.
-    """
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    net.links_df["my_ad_hoc_field"] = 22.5
-
-    print("Network with field...\n ", net.links_df["my_ad_hoc_field"][0:5])
-
-    assert net.links_df["my_ad_hoc_field"][0] == 22.5
-
-def test_add_managed_lane_complex(request, stpaul_net, stpaul_ex_dir, scratch_dir):
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    print("Reading project card ...")
-    project_card_name = "broken_parallel_managed_lane.yml"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path)
-    print("Selecting roadway features ...")
-    selected_link_indices = net.select_roadway_features(project_card.facility)
-
-    attributes_to_update = [p["property"] for p in project_card.properties]
-
-    orig_links = net.links_df.loc[
-        selected_link_indices, net.links_df.columns.intersection(attributes_to_update)
-    ]
-    print("Original Links:\n", orig_links)
-
-    net = net.apply_managed_lane_feature_change(
-        net.select_roadway_features(project_card.facility), project_card.properties
-    )
-
-    rev_links = net.links_df.loc[selected_link_indices, attributes_to_update]
-    print("Revised Links:\n", rev_links)
-
-    net.write(filename="test_ml", path=scratch_dir)
-
-    print("--Finished:", request.node.name)
-
-
-def test_managed_lane_change_functionality(request, stpaul_net, stpaul_ex_dir, scratch_dir):
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    print("Reading project card ...")
-    project_card_name = "test_managed_lanes_change_keyword.yml"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path, validate=False)
-
-    print("Selecting roadway features ...")
-    selected_link_indices = net.select_roadway_features(project_card.facility)
-
-    attributes_to_update = [p["property"] for p in project_card.properties]
-
-    orig_links = net.links_df.loc[
-        selected_link_indices, net.links_df.columns.intersection(attributes_to_update)
-    ]
-    print("Original Links:\n", orig_links)
-
-    net = net.apply_managed_lane_feature_change(
-        net.select_roadway_features(project_card.facility), project_card.properties
-    )
-
-    rev_links = net.links_df.loc[selected_link_indices, attributes_to_update]
-    print("Revised Links:\n", rev_links)
-
-    rev_links.to_csv(os.path.join(scratch_dir, "ml_links.csv"), index=False)
-
-    print("--Finished:", request.node.name)
-
-
-def test_add_default_value(request, stpaul_net, stpaul_ex_dir):
-    """
-    Makes sure we can add a new field with a default value
-    """
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    project_card_name = "select_all_project_card.yml"
-
-    print("Reading project card", project_card_name, "...")
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path)
-
-    print("Selecting roadway features ...")
-    selected_link_indices = net.select_roadway_features(project_card.facility)
-
-    attributes_to_update = [p["property"] for p in project_card.properties]
-
-    net = net.apply_roadway_feature_change(
-        net.select_roadway_features(project_card.facility), project_card.properties
-    )
-
-    print(net.links_df["my_ad_hoc_field_float"].value_counts())
-    print(net.links_df["my_ad_hoc_field_integer"].value_counts())
-    print(net.links_df["my_ad_hoc_field_string"].value_counts())
-
-    assert (
-        net.links_df.loc[net.links_df["my_ad_hoc_field_float"] == 1.2345].shape
-        == net.links_df.shape
-    )
-    assert (
-        net.links_df.loc[net.links_df["my_ad_hoc_field_integer"] == 10].shape
-        == net.links_df.shape
-    )
-    assert (
-        net.links_df.loc[
-            net.links_df["my_ad_hoc_field_string"] == "default value"
-        ].shape
-        == net.links_df.shape
-    )
-
-    print("--Finished:", request.node.name)
-
-def test_add_adhoc_field_from_card(request, stpaul_net, stpaul_ex_dir):
-    """
-    Makes sure new fields can be added from a project card and that
-    they will be the right type.
-    """
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-    project_card_name = "new_fields_project_card.yml"
-
-    print("Reading project card", project_card_name, "...")
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path)
-
-    print("Selecting roadway features ...")
-    selected_link_indices = net.select_roadway_features(project_card.facility)
-
-    attributes_to_update = [p["property"] for p in project_card.properties]
-
-    net = net.apply_roadway_feature_change(
-        net.select_roadway_features(project_card.facility), project_card.properties
-    )
-
-    rev_links = net.links_df.loc[selected_link_indices, attributes_to_update]
-    rev_types = [(a, net.links_df[a].dtypes) for a in attributes_to_update]
-    # rev_types = net.links_df[[attributes_to_update]].dtypes
-    print("Revised Links:\n", rev_links, "\nNew Property Types:\n", rev_types)
-
-    assert net.links_df.loc[selected_link_indices[0], "my_ad_hoc_field_float"] == 1.1
-    assert net.links_df.loc[selected_link_indices[0], "my_ad_hoc_field_integer"] == 2
-    assert (
-        net.links_df.loc[selected_link_indices[0], "my_ad_hoc_field_string"] == "three"
-    )
-    print("--Finished:", request.node.name)
-
-def test_bad_properties_statements(request, small_net):
-    """
-    Makes sure new fields can be added from a project card and that
-    they will be the right type.
-    """
-
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(small_net)
-    ok_properties_change = [{"property": "lanes", "change": 1}]
-    bad_properties_change = [{"property": "my_random_var", "change": 1}]
-    bad_properties_existing = [{"property": "my_random_var", "existing": 1}]
-
-    with pytest.raises(ValueError):
-        net.validate_properties(bad_properties_change)
-
-    with pytest.raises(ValueError):
-        net.validate_properties(ok_properties_change, require_existing_for_change=True)
-
-    with pytest.raises(ValueError):
-        net.validate_properties(bad_properties_existing, ignore_existing=False)
-
-    print("--Finished:", request.node.name)
-
-
-def test_add_delete_roadway_project_card(request, stpaul_net, stpaul_ex_dir):
-    print("\n--Starting:", request.node.name)
-
-    print("Reading network ...")
-    net = copy.deepcopy(stpaul_net)
-    project_cards_list = [
-        "10_simple_roadway_add_change.yml",
-        "11_multiple_roadway_add_and_delete_change.yml",
-    ]
-
-    for card_name in project_cards_list:
-        print("Applying project card - ", card_name, "...")
-        project_card_path = os.path.join(stpaul_ex_dir, "project_cards", card_name)
-        project_card = ProjectCard.read(project_card_path, validate=False)
-
-        orig_links_count = len(net.links_df)
-        orig_nodes_count = len(net.nodes_df)
-
-        missing_nodes = []
-        missing_links = []
-
-        project_card_dictionary = project_card.__dict__
-
-        def _get_missing_nodes(project_card_dictionary):
-            if project_card_dictionary[
-                "category"
-            ].lower() == "roadway deletion" and project_card_dictionary.get("nodes"):
-                for key, val in project_card_dictionary["nodes"].items():
-                    return [v for v in val if v not in net.nodes_df[key].tolist()]
-
-        def _get_missing_links(project_card_dictionary):
-            if project_card_dictionary[
-                "category"
-            ].lower() == "roadway deletion" and project_card_dictionary.get("links"):
-                for key, val in project_card_dictionary["links"].items():
-                    return [v for v in val if v not in net.links_df[key].tolist()]
-
-        # count nodes that are in original network that should be deleted
-        if not project_card_dictionary.get("changes"):
-            m_n = _get_missing_nodes(project_card_dictionary)
-            if m_n:
-                missing_nodes += m_n
-
-            m_l = _get_missing_links(project_card_dictionary)
-            if m_l:
-                missing_links += m_l
-        else:
-            for project_dictionary in project_card_dictionary["changes"]:
-                m_n = _get_missing_nodes(project_dictionary)
-                if m_n:
-                    missing_nodes += m_n
-
-                m_l = _get_missing_links(project_dictionary)
-                if m_l:
-                    missing_links += m_l
-
-        net = net.apply(project_card.__dict__)
-
-        rev_links_count = len(net.links_df)
-        rev_nodes_count = len(net.nodes_df)
-
-        def _count_add_or_delete_features(project_card_dictionary):
-            _links_added = 0
-            _nodes_added = 0
-            _links_deleted = 0
-            _nodes_deleted = 0
-
-            if project_card_dictionary["category"].lower() == "add new roadway":
-                if project_card_dictionary.get("links") is not None:
-                    _links_added = len(project_card_dictionary["links"])
-                if project_card_dictionary.get("nodes") is not None:
-                    _nodes_added = len(project_card_dictionary["nodes"])
-
-            if project_card_dictionary["category"].lower() == "roadway deletion":
-                if project_card_dictionary.get("links") is not None:
-                    print("links", project_card_dictionary["links"])
-                    _links_deleted = sum(
-                        len(project_card_dictionary["links"][key])
-                        for key in project_card_dictionary["links"]
-                    )
-                if project_card_dictionary.get("nodes"):
-                    print("nodes", project_card_dictionary["nodes"])
-                    _nodes_deleted = sum(
-                        len(project_card_dictionary["nodes"][key])
-                        for key in project_card_dictionary["nodes"]
-                    )
-                    print("nodes_deleted:", _nodes_deleted)
-                    print("NODES:", project_card_dictionary["nodes"])
-
-            return {
-                "links_added": _links_added,
-                "nodes_added": _nodes_added,
-                "links_deleted": _links_deleted,
-                "nodes_deleted": _nodes_deleted,
-            }
-
-        links_added = 0
-        links_deleted = 0
-        nodes_added = 0
-        nodes_deleted = 0
-
-        if not project_card_dictionary.get("changes"):
-            count_info = _count_add_or_delete_features(project_card_dictionary)
-            links_added = count_info["links_added"]
-            links_deleted = count_info["links_deleted"]
-            nodes_added = count_info["nodes_added"]
-            nodes_deleted = count_info["nodes_deleted"]
-        else:
-            for project_dictionary in project_card_dictionary["changes"]:
-                count_info = _count_add_or_delete_features(project_dictionary)
-                links_added += count_info["links_added"]
-                links_deleted += count_info["links_deleted"]
-                nodes_added += count_info["nodes_added"]
-                nodes_deleted += count_info["nodes_deleted"]
-
-        num_missing_nodes = len(set(missing_nodes))
-        print("MISSING NODES:", num_missing_nodes)
-
-        num_missing_links = len(set(missing_links))
-        print("MISSING LINKS:", num_missing_links)
-
-        net_links_network = rev_links_count - orig_links_count
-        net_links_project_card = links_added - links_deleted + num_missing_links
-
-        net_nodes_network = rev_nodes_count - orig_nodes_count
-        net_nodes_project_card = nodes_added - nodes_deleted + num_missing_nodes
-        assert net_links_network == net_links_project_card
-        assert net_nodes_network == net_nodes_project_card
-
-    print("--Finished:", request.node.name)
-
-
-@pytest.mark.xfail(strict=True)
-def test_add_roadway_links(request, stpaul_net, stpaul_ex_dir):
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(stpaul_net)
-
-    print("Reading project card ...")
-    # project_card_name = "10_simple_roadway_add_change.yml"
-    project_card_name = "10a_incorrect_roadway_add_change.yml"
-
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path)
-
-    project_card_dictionary = project_card.__dict__
-
-    net.add_new_roadway_feature_change(
-        project_card_dictionary.get("links"), project_card_dictionary.get("nodes")
-    )
-
-    print("--Finished:", request.node.name)
-
-
-def test_existing_managed_lane_apply(request, stpaul_net, stpaul_ex_dir):
-    print("\n--Starting:", request.node.name)
-
-    print("Reading network ...")
-
-    net = copy.deepcopy(stpaul_net)
-
-    print("Reading project card ...")
-    project_card_name = "4_simple_managed_lane.yml"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path)
-
-    print("Selecting roadway features ...")
-    selected_link_indices = net.select_roadway_features(project_card.facility)
-
-    if "managed" in net.links_df.columns:
-        existing_ml_links = len((net.links_df[net.links_df["managed"] == 1]).index)
-    else:
-        existing_ml_links = 0
-
-    print("Existing # of ML links in the network:", existing_ml_links)
-
-    net = net.apply_managed_lane_feature_change(
-        net.select_roadway_features(project_card.facility), project_card.properties
-    )
-
-    new_ml_links = len((net.links_df[net.links_df["managed"] == 1]).index)
-    print("New # of ML links in the network:", new_ml_links)
-
-    assert new_ml_links == existing_ml_links + len(selected_link_indices)
-
-    print("--Finished:", request.node.name)
-
-
-def test_delete_roadway_shape(request, stpaul_net, stpaul_ex_dir):
-    print("\n--Starting:", request.node.name)
-
-    print("Reading network ...")
-    net = copy.deepcopy(stpaul_net)
-
-    print("Reading project card ...")
-    project_card_name = "13_simple_roadway_delete_change.yml"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path, validate=False)
-    project_card_dictionary = project_card.__dict__
-
-    orig_links_count = len(net.links_df)
-    orig_shapes_count = len(net.shapes_df)
-
-    net.delete_roadway_feature_change(
-        project_card_dictionary.get("links"), project_card_dictionary.get("nodes")
-    )
-
-    rev_links_count = len(net.links_df)
-    rev_shapes_count = len(net.shapes_df)
-
-    assert (orig_links_count - rev_links_count) == (
-        orig_shapes_count - rev_shapes_count
-    )
-    assert orig_shapes_count > rev_shapes_count
-    assert orig_links_count > rev_links_count
-
-    print("--Finished:", request.node.name)
-
-
-def test_create_default_geometry(request,stpaul_net, stpaul_ex_dir):
-    print("\n--Starting:", request.node.name)
-
-    print("Reading network ...")
-    net = copy.deepcopy(stpaul_net)
-
-    print("Reading project card ...")
-    project_card_name = "10_simple_roadway_add_change.yml"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path, validate=False)
-    project_card_dictionary = project_card.__dict__
-
-    net.add_new_roadway_feature_change(
-        project_card_dictionary.get("links"), project_card_dictionary.get("nodes")
-    )
-
-    links_without_geometry = net.links_df[net.links_df["geometry"] == ""]
-
-    assert len(links_without_geometry) == 0
-
-    print("--Finished:", request.node.name)
-
-
-def test_add_roadway_shape(request, stpaul_net,stpaul_ex_dir):
-    print("\n--Starting:", request.node.name)
-
-    print("Reading network ...")
-    net = copy.deepcopy(stpaul_net)
-
-    print("Reading project card ...")
-    project_card_name = "10_simple_roadway_add_change.yml"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path, validate=False)
-    project_card_dictionary = project_card.__dict__
-
-    orig_links_count = len(net.links_df)
-    orig_shapes_count = len(net.shapes_df)
-
-    net.add_new_roadway_feature_change(
-        project_card_dictionary.get("links"), project_card_dictionary.get("nodes")
-    )
-
-    rev_links_count = len(net.links_df)
-    rev_shapes_count = len(net.shapes_df)
-
-    assert (rev_links_count - orig_links_count) == (
-        rev_shapes_count - orig_shapes_count
-    )
-    assert rev_shapes_count == orig_shapes_count + 2
-    assert rev_links_count == orig_links_count + 2
-
-    print("--Finished:", request.node.name)
-
-
-def test_create_ml_network_shape(request,stpaul_net, stpaul_ex_dir):
-    print("\n--Starting:", request.node.name)
-    from network_wrangler.roadway import create_managed_lane_network
-
-    print("Reading network ...")
-    net = copy.deepcopy(stpaul_net)
-
-    print("Reading project card ...")
-    project_card_name = "4_simple_managed_lane.yml"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path, validate=False)
-    project_card_dictionary = project_card.__dict__
-
-    _orig_links_count = len(net.links_df)
-    _orig_shapes_count = len(net.shapes_df)
-
-    net = net.apply(project_card_dictionary)
-    net = create_managed_lane_network(net)
-
-    base_model_link_ids = project_card.__dict__["facility"]["links"][0]["model_link_id"]
-    ml_model_link_ids = [
-        MANAGED_LANES_LINK_ID_SCALAR + x for x in base_model_link_ids
-    ]
-    access_model_link_ids = [
-        sum(x) + 1 for x in zip(base_model_link_ids, ml_model_link_ids)
-    ]
-    egress_model_link_ids = [
-        sum(x) + 2 for x in zip(base_model_link_ids, ml_model_link_ids)
-    ]
-
-    gp_links = net.m_links_df[net.m_links_df["model_link_id"].isin(base_model_link_ids)]
-    ml_links = net.m_links_df[net.m_links_df["model_link_id"].isin(ml_model_link_ids)]
-    access_links = net.m_links_df[
-        net.m_links_df["model_link_id"].isin(access_model_link_ids)
-    ]
-    egress_links = net.m_links_df[
-        net.m_links_df["model_link_id"].isin(egress_model_link_ids)
-    ]
-
-    _num_added_links = len(net.links_df) - _orig_links_count
-    _num_added_shapes = len(net.shapes_df) - _orig_shapes_count
-
-    ## 1 Num Added links == added shapes
-    assert _num_added_links == _num_added_shapes
-
-    # 2 new ML links, each ML link has 2 more access/egress links for total of 3 links per ML link
-    # total new links for 2 ML links will be 6 (2*3)
-    _display_c = ["model_link_id", "roadway", "A", "B", "shape_id", "name"]
-    assert (
-        len(net.m_links_df[net.m_links_df["model_link_id"].isin(ml_model_link_ids)])
-        == 2
-    ), f"\n***ML Links\n{ml_links[_display_c]}\
-        \n***ML Links\n{ml_links[_display_c]}\
-        \n***GP Links\n{gp_links[_display_c]}"
-
-    assert (
-        len(net.m_links_df[net.m_links_df["model_link_id"].isin(access_model_link_ids)])
-        == 2
-    ), f"\n***Access Links\n{access_links[_display_c]}\
-        \n***ML Links\n{ml_links[_display_c]}\
-        \n***GP Links\n{gp_links[_display_c]}"
-
-    assert (
-        len(net.m_links_df[net.m_links_df["model_link_id"].isin(egress_model_link_ids)])
-        == 2
-    ), f"\n***Egress Links\n{egress_links[_display_c]}\
-        \n***ML Links\n{ml_links[_display_c]}\
-        \n***GP Links\n{gp_links[_display_c]}"
-
-    print("--Finished:", request.node.name)
-
-
-def test_dot_wrangler_roadway(request,stpaul_ex_dir):
-    print("\n--Starting:", request.node.name)
-
-    print("Reading .wrangler project card ...")
-    project_card_name = "add_highway_lanes.wrangler"
-    project_card_path = os.path.join(stpaul_ex_dir, "project_cards", project_card_name)
-    project_card = ProjectCard.read(project_card_path, validate=False)
-    print(project_card)
-    assert (
-        "self.links_df.loc[self.links_df['lanes'] == 4, 'lanes'] = 12"
-        in project_card.pycode
-    )
-
-
-def test_apply_pycode_roadway(request,stpaul_net):
-    print("\n--Starting:", request.node.name)
-
-    print("Reading network ...")
-    net = copy.deepcopy(stpaul_net)
-
-    print("Apply pycode ...")
-    print(
-        "BEFORE CHANGE...\n",
-        net.links_df.loc[net.links_df["lanes"] == 4, ["model_link_id", "lanes"]],
-    )
-    net = net.apply(
-        {
-            "category": "Calculated Roadway",
-            "project": "megaroads",
-            "pycode": "self.links_df.loc[self.links_df['lanes'] == 4, 'lanes'] = 12",
-        }
-    )
-    print(
-        "AFTER CHANGE...\n",
-        net.links_df.loc[net.links_df["lanes"] == 12, ["model_link_id", "lanes"]],
-    )
-
-
-def test_add_nodes(request,small_net):
-    print("\n--Starting:", request.node.name)
-    net = copy.deepcopy(small_net)
-
-    node_properties = {
-        "X": -93.14412,
-        "Y": 44.87497,
-        "bike_node": 1,
-        "drive_node": 1,
-        "transit_node": 0,
-        "walk_node": 1,
-        "model_node_id": 1234567,
-    }
-
-    net = net.apply(
-        {
-            "category": "add new roadway",
-            "project": "test adding a node",
-            "nodes": [node_properties],
-        }
-    )
-
-    print(net.nodes_df.loc[net.nodes_df.model_node_id == 1234567])
-
-    assert 1234567 in net.nodes_df.model_node_id.tolist()
-
-    # should fail when adding a node with a model_node_id that already exists
-    bad_node_properties = node_properties.copy()
-    bad_node_properties["model_node_id"] = (3494,)  # should already be in network
-    try:
-        net = net.apply(
-            {
-                "category": "add new roadway",
-                "project": "test adding a node already in network",
-                "nodes": [bad_node_properties],
-            },
-        )
-    except ValueError:
-        "expected ValueError when adding a node with a model_node_id that already exists"
-
-
-def test_change_node_xy(small_net):
-    """Tests if X and Y property changes from a project card also update the node geometry."""
-    net = copy.deepcopy(small_net)
-
-    _test_link = net.links_df.iloc[0]
-    _test_link_idx = _test_link[RoadwayNetwork.UNIQUE_LINK_KEY]
-    _test_node = net.nodes_df.loc[[_test_link["A"]]].iloc[0]
-    _test_node_idx = _test_node[[RoadwayNetwork.UNIQUE_NODE_KEY]].iloc[0]
-
-    WranglerLogger.debug(f"Node Index: {_test_node_idx}")
-    WranglerLogger.debug(f"Link Index: {_test_link_idx}")
-    WranglerLogger.info(
-        f"Original Node (Index: {_test_node_idx}):\n{net.nodes_df.loc[_test_node_idx]}"
-    )
-
-    facility = {
-        "nodes": [
-            {"model_node_id": [_test_node_idx]},
-        ]
-    }
-    properties = [
-        {"property": "X", "set": -1000},
-        {"property": "Y", "set": 1000000},
-    ]
-
-    net = net.apply(
-        {
-            "category": "Roadway Property Change",
-            "project": "Update node geometry",
-            "facility": facility,
-            "properties": properties,
-        }
-    )
-
-    WranglerLogger.info(f"Updated Node:\n{net.nodes_df.loc[_test_node_idx]}")
-
-    assert net.nodes_df.loc[_test_node_idx].geometry.x == -1000
-
-    assert (
-        net.links_df.loc[net.links_df[RoadwayNetwork.UNIQUE_LINK_KEY] == 224]
-        .geometry[0]
-        .coords[0][0]
-        == -1000
-    )
